Split structs into mutliple files
