+++ internal/structs/encryptedfile.go
@@ -1,0 +1,7 @@
+package structs
+
+type EncryptedFile struct {
+	Kid          KeyvaultObjectId `json:"kid,omitempty"`
+	Data         string           `json:"data,omitempty"`
+	LastModified JTime            `json:"lastmodified,omitempty"`
+}
